refactor: move scheduleAtInterval to @streamr/utils
