+++ packages/network/src/helpers/scheduler.ts
@@ -1,34 +1,4 @@
 /*
- * @param {number} interval - number of milliseconds to wait after a task is completed
- */
-export const scheduleAtInterval = async (
-    task: () => Promise<void>, 
-    interval: number,
-    executeAtStart: boolean
-): Promise<{ stop: () => void }> => {
-    let timer: NodeJS.Timer|undefined
-    let stopRequested = false
-    const scheduleNext = () => {
-        if (!stopRequested) {
-            timer = setTimeout(async () => {
-                await task()
-                scheduleNext()
-            }, interval)
-        }
-    }
-    if (executeAtStart) {
-        await task()
-    }
-    scheduleNext()
-    return {
-        stop: () => {
-            stopRequested = true
-            clearTimeout(timer!)
-        }
-    }
-}
-
-/*
  * @param {number} interval - execute task when UTC timestamp is divisible by the given interval
  *                            e.g. scheduleAtFixedRate(() => {},  24 * 60 * 60 * 1000) triggers the 
  *                            task once a day at 00:00 UTC
