+++ packages/network/tsconfig.jest.json
@@ -15,6 +15,7 @@
   "references": [
     { "path": "../utils/tsconfig.node.json" },
     { "path": "../test-utils/tsconfig.node.json" },
-    { "path": "../protocol/tsconfig.node.json" }
+    { "path": "../protocol/tsconfig.node.json" },
+    { "path": "../network-tracker/tsconfig.node.json" }
   ]
 }
