+++ packages/utils/src/exports.ts
@@ -1,10 +1,12 @@
 import { Logger } from './Logger'
 import { Multimap } from './Multimap'
+import { scheduleAtInterval } from './scheduleAtInterval'
 import { TimeoutError, withTimeout } from './withTimeout'
 
 export {
     Logger,
     Multimap,
     TimeoutError,
+    scheduleAtInterval,
     withTimeout,
 }
